Kraken: Use of prev/next vj
We don't use anymore the journey_pattern_connection to handle
block_id.

After computing a journey_pattern, if the used vj has a next_vj
we store it in a queue.

After computing all the journey_patterns, we look at all the vj
in this queue, and we try to apply them on the labels.